LayoutTests/fast/canvas/canvas-overloads-setFillColor-expected.txt

Test the behavior of CanvasRenderingContext2D.setFillColor() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.setFillColor() threw exception TypeError: Not enough arguments.
PASS ctx.setFillColor('red') is undefined
PASS ctx.setFillColor(0) is undefined
PASS ctx.setFillColor(0, 0) is undefined
PASS ctx.setFillColor(0, 0, 0) threw exception TypeError: Type error.
PASS ctx.setFillColor(0, 0, 0, 0) is undefined
PASS ctx.setFillColor(0, 0, 0, 0, 0) is undefined
PASS ctx.setFillColor(0, 0, 0, 0, 0, 0) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
